# fmt of non-decimal radix untangled