@pascaldekloe 2025-07-18
Interesting indeed @tgross35. The specialized benchmarks we have at the moment (such as `write_u8_min` and `write_u64_max`) won't scale. Think about how many benchmarks we need for `fmt::LowerExp` with edge cases on alignment and precision. I am curious to hear what you think of the `_10ints_` concept in this patch. The idea is to have a balanced mix of common cases.