@rust-log-analyzer 2025-08-12
The job **`x86_64-gnu-llvm-19-3`** failed! Check out the build log: [(web)](https://github.com/rust-lang/rust/actions/runs/16918367729/job/47938019383) [(plain enhanced)](https://triage.rust-lang.org/gha-logs/rust-lang/rust/47938019383) [(plain)](https://github.com/rust-lang/rust/commit/59e4e19eacd15371afe6d849fe71c1e2f9555a2b/checks/47938019383/logs)

<details><summary><i>Click to see the possible cause of the failure (guessed by this bot)</i></summary>

```plain
   Compiling std v0.0.0 (/checkout/library/std)
error[E0614]: type `u8` cannot be dereferenced
   --> library/core/src/fmt/num.rs:331:64
    |
121 | / macro_rules! impl_Display {
122 | |     ($($Signed:ident, $Unsigned:ident),* ; as $T:ident into $fmt_fn:ident) => {
123 | |
124 | |         $(
...   |
331 | |                     offset = ${concat($fmt_fn, _in_buf_small)}(*self as $T, &mut buf.buf);
    | |                                                                ^^^^^ can't be dereferenced
...   |
372 | |     };
373 | | }
    | |_- in this expansion of `impl_Display!`
...
562 |       impl_Display!(i8, u8, i16, u16, i32, u32, i64, u64, isize, usize; as u64 into display_u64);
    |       ------------------------------------------------------------------------------------------ in this macro invocation

error[E0614]: type `u16` cannot be dereferenced
   --> library/core/src/fmt/num.rs:331:64
    |
121 | / macro_rules! impl_Display {
122 | |     ($($Signed:ident, $Unsigned:ident),* ; as $T:ident into $fmt_fn:ident) => {
123 | |
124 | |         $(
...   |
331 | |                     offset = ${concat($fmt_fn, _in_buf_small)}(*self as $T, &mut buf.buf);
    | |                                                                ^^^^^ can't be dereferenced
...   |
372 | |     };
373 | | }
    | |_- in this expansion of `impl_Display!`
...
562 |       impl_Display!(i8, u8, i16, u16, i32, u32, i64, u64, isize, usize; as u64 into display_u64);
    |       ------------------------------------------------------------------------------------------ in this macro invocation

error[E0614]: type `u32` cannot be dereferenced
   --> library/core/src/fmt/num.rs:331:64
    |
121 | / macro_rules! impl_Display {
122 | |     ($($Signed:ident, $Unsigned:ident),* ; as $T:ident into $fmt_fn:ident) => {
123 | |
124 | |         $(
...   |
331 | |                     offset = ${concat($fmt_fn, _in_buf_small)}(*self as $T, &mut buf.buf);
    | |                                                                ^^^^^ can't be dereferenced
...   |
372 | |     };
373 | | }
    | |_- in this expansion of `impl_Display!`
...
562 |       impl_Display!(i8, u8, i16, u16, i32, u32, i64, u64, isize, usize; as u64 into display_u64);
    |       ------------------------------------------------------------------------------------------ in this macro invocation

error[E0614]: type `u64` cannot be dereferenced
   --> library/core/src/fmt/num.rs:331:64
    |
121 | / macro_rules! impl_Display {
122 | |     ($($Signed:ident, $Unsigned:ident),* ; as $T:ident into $fmt_fn:ident) => {
123 | |
124 | |         $(
...   |
331 | |                     offset = ${concat($fmt_fn, _in_buf_small)}(*self as $T, &mut buf.buf);
    | |                                                                ^^^^^ can't be dereferenced
...   |
372 | |     };
373 | | }
    | |_- in this expansion of `impl_Display!`
...
562 |       impl_Display!(i8, u8, i16, u16, i32, u32, i64, u64, isize, usize; as u64 into display_u64);
    |       ------------------------------------------------------------------------------------------ in this macro invocation

error[E0614]: type `usize` cannot be dereferenced
   --> library/core/src/fmt/num.rs:331:64
    |
121 | / macro_rules! impl_Display {
122 | |     ($($Signed:ident, $Unsigned:ident),* ; as $T:ident into $fmt_fn:ident) => {
123 | |
124 | |         $(
...   |
331 | |                     offset = ${concat($fmt_fn, _in_buf_small)}(*self as $T, &mut buf.buf);
    | |                                                                ^^^^^ can't be dereferenced
...   |
372 | |     };
373 | | }
    | |_- in this expansion of `impl_Display!`
...
562 |       impl_Display!(i8, u8, i16, u16, i32, u32, i64, u64, isize, usize; as u64 into display_u64);
    |       ------------------------------------------------------------------------------------------ in this macro invocation

For more information about this error, try `rustc --explain E0614`.
[RUSTC-TIMING] core test:false 24.199
error: could not compile `core` (lib) due to 5 previous errors
```

</details>